Refactor utilities and routes
